sched/fair: Calculate avg_vruntime with additional sched_entity
Add an additional parameter se (sched_entity) to avg_vruntime() and rename it
to __avg_vruntime(). __avg_vruntime() can now calculate avg_vruntime as if se
is on the runqueue.

This can be used by update_entity_lag() to calculate the lag of a sched_entity
if it is not on the runqueue.

Also add an wrapper function avg_vruntime() that accepts original
parameters and calls __avg_vruntime() with se set to NULL to preserve
the original behavior of avg_vruntime().

Signed-off-by: Yan-Jie Wang <yanjiewtw@gmail.com>